// src/lib/database-types.ts
export interface SystemFields {
    id: string;
    access_read: string[] | null;
    access_edit: string[] | null;
    access_full: string[] | null;
    access_deny: string[] | null;
    created_at: string;
    updated_at: string;
    trashed_at: string | null;
    deleted_at: string | null;
}
// ...
export type DbRecord<T extends Record<string, any> = Record<string, any>> = SystemFields & T;
// ...
export interface ModelRecord extends DbRecord {
    model_name: string;
    status: string;
    sudo?: boolean;
    external?: boolean;
}

export interface FieldRecord extends DbRecord {
    model_name: string;
    field_name: string;
    type: string;
    required?: boolean;
    default_value?: any;
    description?: string;
    pattern?: string;
    minimum?: number;
    maximum?: number;
    min_length?: number;
    max_length?: number;
    enum_values?: string[];
    relationship_type?: 'owned' | 'referenced';
    relationship_model?: string;
    relationship_name?: string;
    relationship_field?: string;
    relationship_cascade_delete?: boolean;
    relationship_required?: boolean;
}